# docs/api/openapi-query.yaml
openapi: 3.1.0
info:
  title: siglake query
  description: |-
    DataFusion SQL over the Iceberg tables, index and index-template management, batch jobs, GDPR delete tasks, and a Jaeger-compatible trace read API.

    Ingest lives on a separate server (`openapi-ingest.yaml`) — this one only reads, apart from index management and delete tasks.
  license:
    name: Apache-2.0
  version: 0.1.0
servers:
- url: http://localhost:8089
  description: Default `--bind` address
paths:
...
  /api/v1/index-templates:
    get:
      tags:
      - index-templates
      summary: List every index template in the caller's tenant namespace.
      description: |-
        Templates apply a document mapping to indexes created later whose id matches
        one of their glob patterns; the highest `priority` match wins. Templates are
        never shared across tenant namespaces.
      operationId: list_templates
      responses:
        '200':
          description: All index templates.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IndexTemplate'
        '401':
          description: Missing or invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '403':
          description: Tenant refused before the handler runs. This server derives the tenant from a verified JWT claim when one is configured (`--oidc-tenant-claim`), and this token carries none, or carries one that is not a usable tenant id (`[A-Za-z0-9_-]`, 1..=128 chars). Answered by the auth middleware, so it can reach every operation behind it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '500':
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
  /api/v1/index-templates/{id}:
    put:
      tags:
      - index-templates
      summary: Create or replace an index template.
      description: |-
        Upsert: there is no separate `POST`. The body's `template_id` must equal the
        path id. Templates are applied at index-creation time, so editing one does
        not retroactively change existing indexes. The template is stored only in
        the caller's tenant namespace.
      operationId: put_template
      parameters:
      - name: id
        in: path
        description: Template id. Must match the body's `template_id`.
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndexTemplate'
        required: true
      responses:
        '200':
          description: Stored. Body echoes the template.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndexTemplate'
        '400':
          description: Invalid template, or the path id does not match the body's `template_id`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '401':
          description: Missing or invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '403':
          description: Tenant refused before the handler runs. This server derives the tenant from a verified JWT claim when one is configured (`--oidc-tenant-claim`), and this token carries none, or carries one that is not a usable tenant id (`[A-Za-z0-9_-]`, 1..=128 chars). Answered by the auth middleware, so it can reach every operation behind it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '500':
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
    delete:
      tags:
      - index-templates
      summary: Delete an index template from the caller's tenant namespace.
      operationId: delete_template
      parameters:
      - name: id
        in: path
        description: Template id.
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Deleted. Empty body.
        '401':
          description: Missing or invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '403':
          description: Tenant refused before the handler runs. This server derives the tenant from a verified JWT claim when one is configured (`--oidc-tenant-claim`), and this token carries none, or carries one that is not a usable tenant id (`[A-Za-z0-9_-]`, 1..=128 chars). Answered by the auth middleware, so it can reach every operation behind it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '404':
          description: No such template.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '500':
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
  /api/v1/indexes:
    get:
      tags:
      - indexes
      summary: List every index in the caller's tenant.
      operationId: list
      responses:
        '200':
          description: All index configurations.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IndexConfig'
        '401':
          description: Missing or invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '403':
          description: Tenant refused before the handler runs. This server derives the tenant from a verified JWT claim when one is configured (`--oidc-tenant-claim`), and this token carries none, or carries one that is not a usable tenant id (`[A-Za-z0-9_-]`, 1..=128 chars). Answered by the auth middleware, so it can reach every operation behind it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '500':
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
    post:
      tags:
      - indexes
      summary: Create an index.
      description: |-
        The body is the full index configuration: its document mapping (typed
        columns and their tokenizers), the required timestamp field, and optional
        retention. Attributes not declared here are handled per `doc_mapping.mode`.
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndexConfig'
        required: true
      responses:
        '201':
          description: Created. Body is the stored configuration as the catalog normalized it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndexConfig'
        '400':
          description: Invalid configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '401':
          description: Missing or invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '403':
          description: Tenant refused before the handler runs. This server derives the tenant from a verified JWT claim when one is configured (`--oidc-tenant-claim`), and this token carries none, or carries one that is not a usable tenant id (`[A-Za-z0-9_-]`, 1..=128 chars). Answered by the auth middleware, so it can reach every operation behind it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '409':
          description: An index with this id already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '500':
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
  /api/v1/indexes/{id}:
    get:
      tags:
      - indexes
      summary: Fetch one index configuration.
      operationId: get
      parameters:
      - name: id
        in: path
        description: Index id.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: The index configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndexConfig'
        '401':
          description: Missing or invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '403':
          description: Tenant refused before the handler runs. This server derives the tenant from a verified JWT claim when one is configured (`--oidc-tenant-claim`), and this token carries none, or carries one that is not a usable tenant id (`[A-Za-z0-9_-]`, 1..=128 chars). Answered by the auth middleware, so it can reach every operation behind it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '404':
          description: No such index.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '500':
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
    put:
      tags:
      - indexes
      summary: Replace an index configuration.
      description: |-
        Schema changes are additive only — the catalog rejects an update that would
        drop or retype an existing column. The body's `index_id` must equal the path
        id.
      operationId: update
      parameters:
      - name: id
        in: path
        description: Index id. Must match the body's `index_id`.
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndexConfig'
        required: true
      responses:
        '200':
          description: Updated. Body is the stored configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndexConfig'
        '400':
          description: Invalid configuration, or the path id does not match the body's `index_id`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '401':
          description: Missing or invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '403':
          description: Tenant refused before the handler runs. This server derives the tenant from a verified JWT claim when one is configured (`--oidc-tenant-claim`), and this token carries none, or carries one that is not a usable tenant id (`[A-Za-z0-9_-]`, 1..=128 chars). Answered by the auth middleware, so it can reach every operation behind it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '404':
          description: No such index.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '500':
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
    delete:
      tags:
      - indexes
      summary: Delete an index.
      description: |-
        Removes the catalog entry only. Committed data files remain at the table
        location; the current retention and orphan-GC sweeps cannot reclaim them
        after the catalog entry is gone.
      operationId: delete
      parameters:
      - name: id
        in: path
        description: Index id.
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Deleted. Empty body.
        '401':
          description: Missing or invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '403':
          description: Tenant refused before the handler runs. This server derives the tenant from a verified JWT claim when one is configured (`--oidc-tenant-claim`), and this token carries none, or carries one that is not a usable tenant id (`[A-Za-z0-9_-]`, 1..=128 chars). Answered by the auth middleware, so it can reach every operation behind it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '404':
          description: No such index.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '500':
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
...
components:
  schemas:
    ApiErrorBody:
      type: object
      description: |-
        Error body used by every endpoint on this server.

        Spec-only. Some errors additionally flatten diagnostic keys into the top
        level — `cost` on a rejected or timed-out query, and `cost` plus `stats` on
        the mid-flight rows breaker — so this schema deliberately allows additional
        properties rather than pinning an exact shape.
      required:
      - error
      - code
      properties:
        code:
          type: integer
          format: int32
          description: Echo of the HTTP status code.
          minimum: 0
        cost:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/CostReport'
            description: Present when the failure was cost- or time-related.
        error:
          type: string
          description: Human-readable failure description.
...
    CostReport:
      type: object
      description: |-
        Cost report returned by `/explain` and attached to every successful
        query response.
      required:
      - estimated_bytes_scanned
      - estimated_rows_processed
      - estimated_runtime_seconds
      - complexity_class
      - warnings
      - exact
      properties:
        complexity_class:
          $ref: '#/components/schemas/ComplexityClass'
        estimated_bytes_scanned:
          type: integer
          format: int64
          minimum: 0
        estimated_rows_processed:
          type: integer
          format: int64
          minimum: 0
        estimated_runtime_seconds:
          type: number
          format: double
        exact:
          type: boolean
          description: |-
            `true` when the underlying statistics are exact; `false` when
            they fell back to a heuristic.
        files_considered:
          type:
          - integer
          - 'null'
          description: |-
            Live files considered BEFORE the manifest time-bounds prune —
            `files_considered - files_to_scan` is planning-time pruning's win.
          minimum: 0
        files_to_scan:
          type:
          - integer
          - 'null'
          minimum: 0
        warnings:
          type: array
          items:
            type: string
...
    DocMapping:
      type: object
      description: Document mapping for one index.
      required:
      - field_mappings
      - timestamp_field
      properties:
        default_search_fields:
          type: array
          items:
            type: string
          description: Fields searched by default for bare text queries.
        field_mappings:
          type: array
          items:
            $ref: '#/components/schemas/FieldMappingSchema'
          description: Declared typed fields in stable order.
        mode:
          $ref: '#/components/schemas/MappingMode'
          description: Handling mode for unmapped attributes.
        tag_fields:
          type: array
          items:
            type: string
          description: Fields treated as low-cardinality tags for pruning.
        timestamp_field:
          type: string
          description: Required event-time field.
      additionalProperties: false
...
    IndexConfig:
      type: object
      description: Complete index configuration for one logical index.
      required:
      - index_id
      - doc_mapping
      properties:
        doc_mapping:
          $ref: '#/components/schemas/DocMapping'
          description: Declared document mapping.
        index_at_flush:
          type:
          - boolean
          - 'null'
          description: |-
            Whether the ingest drain builds the raw-text search indexes (inverted
            index, trigram + row-group token blooms, Puffin sidecar) inline with
            every flush, or defers them to compaction (files gain indexes when they
            consolidate at L1). Inline makes the freshest data fully searchable at
            ~30 % append-time cost (measured); deferring buys drain throughput and
            is the right default for firehose streams — the query funnel tolerates
            mixed indexed/unindexed files by falling back to a scan. `None` inherits
            the deployment default (`SIGLAKE_INDEX_AT_FLUSH`, inline unless `0`).
        index_id:
          type: string
          description: Human-facing index identifier.
        retention:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/RetentionPolicy'
            description: Optional retention policy.
      additionalProperties: false
    IndexTemplate:
      type: object
      description: Index-template metadata persisted in the warehouse config area.
      required:
      - template_id
      - index_id_patterns
      - priority
      - doc_mapping
      properties:
        doc_mapping:
          $ref: '#/components/schemas/DocMapping'
          description: Document mapping to materialize onto the created index.
        index_id_patterns:
          type: array
          items:
            type: string
          description: Glob patterns matched against candidate index ids.
        priority:
          type: integer
          format: int32
          description: Higher values win when multiple templates match.
        retention:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/RetentionPolicy'
            description: Optional retention policy copied into the created index.
        template_id:
          type: string
          description: Stable template identifier.
      additionalProperties: false
...
    MappingMode:
      type: string
      description: Behavior for attributes that are not declared in `field_mappings`.
      enum:
      - dynamic
      - lenient
      - strict
...
    RetentionPolicy:
      type: object
      description: Retention configuration for one index.
      required:
      - period_secs
      properties:
        period_secs:
          type: integer
          format: int64
          description: Retention horizon in seconds.
          minimum: 0
        schedule:
          type:
          - string
          - 'null'
          description: Optional evaluator schedule string.
      additionalProperties: false
...
  securitySchemes:
    query_token:
      type: http
      scheme: bearer
      description: '`Authorization: Bearer <token>`. With OIDC configured the token is a JWT and its claims select the tenant; otherwise it is matched against the cluster''s static token list. A server started with neither accepts unauthenticated requests.'
security:
- query_token: []
...